# aula 06/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Filmes DS2M - SENAI JANDIRA
  description: API destinada ao controle de filmes, personagens, gêneros, nacionalidades e sexo
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
paths:
...
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo filme
      description: Adiciona um novo filme no banco de dados
      requestBody:
        description: JSON contendo os dados do novo filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeInput'
        required: true
      responses:
        '201':
          description: Filme criado com sucesso
          content:
            
        '400':
          description: Dados inválidos
      
        '415':
          description: Content-Type inválido
        
        '500':
          description: Erro interno no servidor
         
    get:
      tags:
        - Filme
      summary: Lista todos os filmes
      description: Retorna todos os filmes registrados no sistema
      responses:
        '200':
          description: Listagem realizada com sucesso
          
        '404':
          description: Nenhum filme encontrado
         
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um filme pelo ID
      description: Retorna um único filme com base no ID informado
      parameters:
        - name: id
          in: path
          description: ID do filme a ser buscado
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Filme encontrado
          
        '400':
          description: ID inválido
          
        '404':
          description: Filme não encontrado
          
        '500':
          description: Erro interno no servidor
         

    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo ID
      description: Localiza e atualiza os dados de um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID do filme a ser atualizado
          required: true
          schema:
            type: integer
      requestBody:
        description: Dados atualizados do filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeInput'
        required: true
      responses:
        '200':
          description: Filme atualizado com sucesso
      
        '400':
          description: Dados ou ID inválidos

        '404':
          description: Filme não encontrado
         
        '415':
          description: Content-Type inválido
          
        '500':
          description: Erro interno no servidor
       
    delete:
      tags:
        - Filme
      summary: Exclui um filme pelo ID
      description: Remove um filme do banco de dados pelo ID
      parameters:
        - name: id
          in: path
          description: ID do filme a ser excluído
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Filme excluído com sucesso
        
        '400':
          description: ID inválido
        
        '404':
          description: Filme não encontrado
         
        '500':
          description: Erro interno no servidor
    

  # ==================== SEXO ====================
  /v1/senai/sexo:
    post:
      tags:
        - Sexo
      summary: Adiciona um novo sexo
      description: Adiciona um novo registro de sexo no banco de dados
      requestBody:
        description: JSON contendo os dados do sexo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SexoInput'
        required: true
      responses:
        '201':
          description: Sexo criado com sucesso
          
        '400':
          description: Dados inválidos
         
        '415':
          description: Content-Type inválido
       
        '500':
          description: Erro interno no servidor
       

    get:
      tags:
        - Sexo
      summary: Lista todos os sexos
      description: Retorna todos os registros de sexo cadastrados
      responses:
        '200':
          description: Listagem realizada com sucesso
         
        '404':
          description: Nenhum registro encontrado
         
        '500':
          description: Erro interno no servidor
  /v1/senai/sexo/{id}:
    get:
      tags:
        - Sexo
      summary: Busca um sexo pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Registro encontrado
         
        '400':
          description: ID inválido
          
        '404':
          description: Registro não encontrado
         
        '500':
          description: Erro interno no servidor
          

    put:
      tags:
        - Sexo
      summary: Atualiza um sexo pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SexoInput'
        required: true
      responses:
        '200':
          description: Atualizado com sucesso

        '400':
          description: Dados ou ID inválidos
        
        '404':
          description: Registro não encontrado
         
        '415':
          description: Content-Type inválido
         
        '500':
          description: Erro interno no servidor
        

    delete:
      tags:
        - Sexo
      summary: Exclui um sexo pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Excluído com sucesso
     
        '400':
          description: ID inválido
         
        '404':
          description: Registro não encontrado
          
        '500':
          description: Erro interno no servidor
        


  # ==================== PERSONAGEM ====================
  /v1/senai/personagem:
    post:
      tags:
        - Personagem
      summary: Adiciona um novo personagem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonagemInput'
        required: true
      responses:
        '201':
          description: Personagem criado com sucesso
          
        '400':
          description: Dados inválidos
         
        '415':
          description: Content-Type inválido
          
        '500':
          description: Erro interno no servidor

    get:
      tags:
        - Personagem
      summary: Lista todos os personagens
      responses:
        '200':
          description: Listagem realizada com sucesso

        '404':
          description: Nenhum personagem encontrado

        '500':
          description: Erro interno no servidor
  /v1/senai/personagem/{id}:
    get:
      tags:
        - Personagem
      summary: Busca um personagem pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Personagem encontrado
      
        '400':
          description: ID inválido
    
        '404':
          description: Personagem não encontrado
         
        '500':
          description: Erro interno no servidor
         
    put:
      tags:
        - Personagem
      summary: Atualiza um personagem pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonagemInput'
        required: true
      responses:
        '200':
          description: Personagem atualizado com sucesso
    
        '400':
          description: Dados ou ID inválidos
       
        '404':
          description: Personagem não encontrado
          
        '415':
          description: Content-Type inválido
          
        '500':
          description: Erro interno no servidor
          

    delete:
      tags:
        - Personagem
      summary: Exclui um personagem pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Personagem excluído com sucesso
          
        '400':
          description: ID inválido
          
        '404':
          description: Personagem não encontrado
          
        '500':
          description: Erro interno no servidor
         


  # ==================== GÊNERO ====================
  /v1/senai/genero:
    post:
      tags:
        - Gênero
      summary: Adiciona um novo gênero
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneroInput'
        required: true
      responses:
        '201':
          description: Gênero criado com sucesso
        
        '400':
          description: Dados inválidos
        
        '415':
          description: Content-Type inválido
        
        '500':
          description: Erro interno no servidor
         

    get:
      tags:
        - Gênero
      summary: Lista todos os gêneros
      responses:
        '200':
          description: Listagem realizada com sucesso
          
        '404':
          description: Nenhum gênero encontrado
        
        '500':
          description: Erro interno no servidor
  /v1/senai/genero/{id}:
    get:
      tags:
        - Gênero
      summary: Busca um gênero pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero encontrado
         
        '400':
          description: ID inválido
          
        '404':
          description: Gênero não encontrado
          
        '500':
          description: Erro interno no servidor
          

    put:
      tags:
        - Gênero
      summary: Atualiza um gênero pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneroInput'
        required: true
      responses:
        '200':
          description: Gênero atualizado com sucesso
          
           
        '400':
          description: Dados ou ID inválidos
          
        '404':
          description: Gênero não encontrado
         
        '415':
          description: Content-Type inválido
          
        '500':
          description: Erro interno no servidor
         

    delete:
      tags:
        - Gênero
      summary: Exclui um gênero pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero excluído com sucesso
          
        '400':
          description: ID inválido
  
        '404':
          description: Gênero não encontrado
     
        '500':
          description: Erro interno no servidor
      


  # ==================== NACIONALIDADE ====================
  /v1/senai/nacionalidade:
    post:
      tags:
        - Nacionalidade
      summary: Adiciona uma nova nacionalidade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NacionalidadeInput'
        required: true
      responses:
        '201':
          description: Nacionalidade criada com sucesso

        '400':
          description: Dados inválidos
      
        '415':
          description: Content-Type inválido
    
        '500':
          description: Erro interno no servidor
        

    get:
      tags:
        - Nacionalidade
      summary: Lista todas as nacionalidades
      responses:
        '200':
          description: Listagem realizada com sucesso
         
        '404':
          description: Nenhuma nacionalidade encontrada
         
        '500':
          description: Erro interno no servidor
  /v1/senai/nacionalidade/{id}:
    get:
      tags:
        - Nacionalidade
      summary: Busca uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Nacionalidade encontrada

        '400':
          description: ID inválido
         
        '404':
          description: Nacionalidade não encontrada
        
        '500':
          description: Erro interno no servidor
         

    put:
      tags:
        - Nacionalidade
      summary: Atualiza uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NacionalidadeInput'
        required: true
      responses:
        '200':
          description: Nacionalidade atualizada com sucesso
   
        '400':
          description: Dados ou ID inválidos
        
        '404':
          description: Nacionalidade não encontrada
       
        '415':
          description: Content-Type inválido
         
        '500':
          description: Erro interno no servidor
       

    delete:
      tags:
        - Nacionalidade
      summary: Exclui uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Nacionalidade excluída com sucesso
      
        '400':
          description: ID inválido
    
        '404':
          description: Nacionalidade não encontrada
   
        '500':
          description: Erro interno no servidor

  # ==================== ATIVIDADE ====================
  /v1/senai/atividade:
    post:
      tags:
        - Atividade
      summary: Adiciona uma nova atividade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtividadeInput'
        required: true
      responses:
        '201':
          description: Atividade criada com sucesso

        '400':
          description: Dados inválidos
      
        '415':
          description: Content-Type inválido
    
        '500':
          description: Erro interno no servidor
        

    get:
      tags:
        - Atividade
      summary: Lista todas as Atividades
      responses:
        '200':
          description: Listagem realizada com sucesso
         
        '404':
          description: Nenhuma Atividade encontrada
         
        '500':
          description: Erro interno no servidor
  /v1/senai/atividade/{id}:
    get:
      tags:
        - Atividade
      summary: Busca uma atividade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Atividade encontrada

        '400':
          description: ID inválido
         
        '404':
          description: Atividade não encontrada
        
        '500':
          description: Erro interno no servidor
         

    put:
      tags:
        - Atividade
      summary: Atualiza uma atividade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtividadeInput'
        required: true
      responses:
        '200':
          description: Atividade atualizada com sucesso
   
        '400':
          description: Dados ou ID inválidos
        
        '404':
          description: Atividade não encontrada
       
        '415':
          description: Content-Type inválido
         
        '500':
          description: Erro interno no servidor
       

    delete:
      tags:
        - Atividade
      summary: Exclui uma atividade pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Atividade excluída com sucesso
      
        '400':
          description: ID inválido
    
        '404':
          description: Atividade não encontrada
   
        '500':
          description: Erro interno no servidor



  # ==================== CLASSIFICACAO ====================
  /v1/senai/classificacao:
    post:
      tags:
        - Classificação
      summary: Adiciona uma nova classificação
      description: Adiciona uma nova classificação indicativa no banco de dados
      requestBody:
        description: JSON contendo os dados da nova classificação
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificacaoInput'
        required: true
      responses:
        '201':
          description: Classificação criada com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    get:
      tags:
        - Classificação
      summary: Lista todas as classificações
      description: Retorna todas as classificações cadastradas no sistema
      responses:
        '200':
          description: Listagem realizada com sucesso
        '404':
          description: Nenhuma classificação encontrada
        '500':
          description: Erro interno no servidor
...
  /v1/senai/pessoa:
    post:
      tags:
        - Pessoa
      summary: Adiciona uma nova pessoa
      description: Adiciona uma nova pessoa no banco de dados
      requestBody:
        description: JSON contendo os dados da nova pessoa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PessoaInput'
        required: true
      responses:
        '201':
          description: Pessoa criada com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    get:
      tags:
        - Pessoa
      summary: Lista todas as pessoas
      description: Retorna todas as pessoas cadastradas no sistema
      responses:
        '200':
          description: Listagem realizada com sucesso
        '404':
          description: Nenhuma pessoa encontrada
        '500':
          description: Erro interno no servidor
...
components:
  schemas:
    FilmeInput:
      required:
        - nome
        - data_lancamento
        - duracao
        - sinopse
        - valor
        - capa
        - id_classificacao
      type: object
      properties:
        nome:
          type: string
          maxLength: 80
          example: Matrix
        data_lancamento:
          type: string
          example: '1999-03-31'
        duracao:
          type: string
          example: '02:16:00'
        sinopse:
          type: string
          example: Um hacker descobre a verdade sobre sua realidade
        avaliacao:
          type: number
          example: 8.7
        valor:
          type: number
          example: 19.90
        capa:
          type: string
          maxLength: 255
          example: 'https://exemplo.com/capa.jpg'
        id_classificacao:
          type: integer
          example: 1
    SexoInput:
      required:
        - nome
        - sigla
      type: object
      properties:
        nome:
          type: string
          maxLength: 40
          example: Masculino
        sigla:
          type: string
          maxLength: 4
          example: M
    PersonagemInput:
      required:
        - nome
      type: object
      properties:
        nome:
          type: string
          example: Cirilo
    GeneroInput:
      required:
        - nome
      type: object
      properties:
        nome:
          type: string
          example: Ação
    NacionalidadeInput:
      required:
        - nome
      type: object
      properties:
        nome:
          type: string
          maxLength: 50
          example: Brasileiro
    AtividadeInput:
      required:
        - nome
      type: object
      properties:
        nome:
          type: string
          maxLength: 45
          example: Ator
    ClassificacaoInput:
      required:
        - nome
        - sigla
        - descricao
      type: object
      properties:
        nome:
          type: string
          maxLength: 50
          example: Livre
        sigla:
          type: string
          maxLength: 5
          example: L
        descricao:
          type: string
          maxLength: 200
          example: Conteúdo adequado para todas as idades
...